features reward

+++ database/migrations/2024_06_08_125547_create_wastes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wastes', function (Blueprint $table) {
             $table->id();
             $table->string('jenis_sampah');
-            $table->integer('poin');
+            $table->float('poin');
             $table->string('image');
             $table->timestamps();
         });
